Improve LED handling
If LED state is currently LED_STATE_OFF or LED_STATE_ON we
leave the LED where it is (and don't restart the standby timer).
If LED state is currently in LED_STATE_API we
turn the LED on (which brings it in LED_STATE_ON) and restarts the standby timer.